Change place of Task validation from model to form (small modification in form). Update todoapp:view to use the new form. (add authentication Part 5)

+++ todoapp/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from django.http import Http404, HttpResponseRedirect
+from django.http import  Http404, HttpResponseRedirect, HttpResponseBadRequest
 from django.urls import reverse
 from django.shortcuts import get_object_or_404
 
@@ -29,27 +29,35 @@ def tasks(request):
             }
         return render(request, 'todoapp/tasks.html', context)
 
+    # create new task
     if request.method == 'POST':
-        form = TaskForm(data=request.POST)
+        form = TaskForm(user=request.user, data=request.POST)
         if form.is_valid():
             new_task = form.save(commit=False)
             new_task.owner = request.user
             new_task.save()
             form.save_m2m()
             return HttpResponseRedirect(reverse('todoapp:tasks'))
+        else: # TODO: improve it
+            return HttpResponseBadRequest()
 
 @login_required
 def hashtags(request):
     if request.method == 'GET':
-        hashtags = Hashtag.objects.all()
+        hashtags = Hashtag.objects.filter(owner=request.user)
         context = {'hashtags': hashtags}
         return render(request, 'todoapp/hashtags.html', context)
 
+    # create new hashtag
     if request.method == 'POST':
         form = HashtagForm(data=request.POST)
         if form.is_valid():
-            new_hashtag = form.save(commit=True)
+            new_hashtag = form.save(commit=False)
+            new_hashtag.owner = request.user
+            new_hashtag.save()
             return HttpResponseRedirect(reverse('todoapp:hashtags'))
+        else: # TODO: improve it
+            return HttpResponseBadRequest()
 
 @login_required
 def task(request, task_id):
@@ -60,29 +68,34 @@ def task(request, task_id):
 
     if request.method == 'GET':
         tick_task_form = TickTaskForm(instance=task)
+        context = {'task': task, 'tick_task_form': tick_task_form}
+        return render(request, 'todoapp/task.html', context)
 
     if request.method == 'POST':
-
         # user want tick on/off task
         if 'tick_task_form' in request.POST: 
             task.completed = not task.completed
             task.save() # Should I check form?
             return HttpResponseRedirect(reverse('todoapp:task',
                                         args=[task_id]))
-        else: # user want to create new task
-            form = TaskForm(instance=task, data=request.POST)
+        else: # user want to update task
+            form = TaskForm(user=request.user, instance=task, data=request.POST)
             if form.is_valid():
                 form.save()
                 return HttpResponseRedirect(reverse('todoapp:task',
                                             args=[task_id]))
+            else: # TODO: improve it
+                return HttpResponseBadRequest()
 
-    context = {'task': task, 'tick_task_form': tick_task_form}
-    return render(request, 'todoapp/task.html', context)
+
 
 @login_required
 def hashtag(request, hashtag_id):
     hashtag = get_object_or_404(Hashtag, id=hashtag_id)
     tasks_in_which_used = hashtag.task_set.all()
+
+    if hashtag.owner != request.user:
+        raise Http404
 
     if request.method == 'GET':
         context = {
@@ -98,10 +111,12 @@ def hashtag(request, hashtag_id):
             form.save()
             return HttpResponseRedirect(reverse('todoapp:hashtag',
                                         args=[hashtag_id]))
+        else: # TODO: improve it
+            return HttpResponseBadRequest()
 
 @login_required
 def new_task(request):
-    form = TaskForm()
+    form = TaskForm(user=request.user)
     content = {'form': form}
     return render(request, 'todoapp/new_task.html', content)
 
@@ -118,13 +133,16 @@ def edit_task(request, task_id):
     if task.owner != request.user:
         raise Http404
 
-    form = TaskForm(instance=task)
+    form = TaskForm(user=request.user, instance=task)
     content = {'task': task, 'form': form}
     return render(request, 'todoapp/edit_task.html', content)
 
 @login_required
 def edit_hashtag(request, hashtag_id):
     hashtag = get_object_or_404(Hashtag, id=hashtag_id)
+
+    if hashtag.owner != request.user:
+        raise Http404
 
     form = HashtagForm(instance=hashtag)
     content = {'hashtag': hashtag, 'form': form}
@@ -133,6 +151,9 @@ def edit_hashtag(request, hashtag_id):
 @login_required
 def delete_task(request, task_id):
     task = get_object_or_404(Task, id=task_id)
+
+    if task.owner != request.user:
+        raise Http404
 
     if request.method == 'GET':
         form = DeleteTaskForm(instance=task)
@@ -145,10 +166,15 @@ def delete_task(request, task_id):
         if form.is_valid():
             task.delete()
             return HttpResponseRedirect(reverse('todoapp:tasks'))
+        else: # TODO: improve it
+            return HttpResponseBadRequest()
 
 @login_required
 def delete_hashtag(request, hashtag_id):
     hashtag = get_object_or_404(Hashtag, id=hashtag_id)
+
+    if hashtag.owner != request.user:
+        raise Http404
 
     if request.method == 'GET':
         form = DeleteHashtagForm(instance=hashtag)
@@ -161,3 +187,5 @@ def delete_hashtag(request, hashtag_id):
         if form.is_valid():
             hashtag.delete()
             return HttpResponseRedirect(reverse('todoapp:hashtags'))
+        else: # TODO: improve it
+            return HttpResponseBadRequest()
